+++ lib/data.ml
@@ -29,15 +29,15 @@ type 'a t =
   | Nil
   | Array of 'a t array
   | Dict of 'a t Map.String.t
-  | Const of Const.t * Ty.Variant.extra
+  | Const of Const.t
 
 let other x = Other x
 let null = Nil
-let const x e = Const (x, e)
-let int i = Const (Int i, Not_bool)
-let bool i = Const (Int i, Bool)
-let string s = Const (String s, Not_bool)
-let float f = Const (Float f, Not_bool)
+let const x = Const x
+let int i = Const (Int i)
+let bool i = Const (Int i)
+let string s = Const (String s)
+let float f = Const (Float f)
 let some x = Array [| x |]
 let dict m = Dict m
 let tuple a = Array a
@@ -57,17 +57,13 @@ let rec flat_map f = function
   | Nil -> Nil
   | Array a -> Array (Array.map (flat_map f) a)
   | Dict d -> Dict (Map.String.map (flat_map f) d)
-  | Const (a, b) -> Const (a, b)
+  | Const a -> Const a
 
-let get_const = function Const (x, _) -> x | _ -> assert false
+let get_const = function Const x -> x | _ -> assert false
 let get_tuple = function Array t -> t | _ -> assert false
 let get_dict = function Dict t -> t | _ -> assert false
 let is_null = function Nil -> true | _ -> false
-
-let get_nullable = function
-  | Nil -> None
-  | Array [| t |] -> Some t
-  | _ -> assert false
+let get_nullable = function Array [| t |] -> Some t | _ -> None
 
 let fold_list f acc l =
   let rec aux i acc = function
@@ -81,12 +77,4 @@ let fold_list f acc l =
 
 let fold_dict f acc = function
   | Dict m -> Map.String.fold (fun k v acc -> f ~index:(string k) acc v) m acc
-  | _ -> assert false
-
-let echo = function
-  | Const (Int 0, Bool) -> "false"
-  | Const (_, Bool) -> "true"
-  | Const (Int i, _) -> string_of_int i
-  | Const (String s, _) -> s
-  | Const (Float n, _) -> Printf.sprintf "%g" n
   | _ -> assert false