+++ test/data_test.ml
@@ -34,7 +34,7 @@ let records () =
   let props_json = {|{"a": {"a": "a", "b": 1, "c": 1.0}}|} in
   let src =
     "{% match a with {a: \"z\", b: 0, c: 0.0} %}\n\
-     {% with {a, b, c} %} {{ a }} {{ b }} {{ c }}\n\
+     {% with {a, b, c} %} {{ a }} {{ %i b }} {{ %g c }}\n\
      {%~ /match %}"
   in
   let result = " a 1 1" in
@@ -95,7 +95,7 @@ let tuples () =
       end)
   in
   let props_json = {|{"a": ["a", 1.0]}|} in
-  let src = "{% match a with (a, b) %} {{ a }} {{ b }} {% /match %}" in
+  let src = "{% match a with (a, b) %} {{ a }} {{ %g b }} {% /match %}" in
   let result = " a 1 " in
   check "List 1 yojoson" result (render_yojson src props_json);
   check "List 1 js" result (render_js src props_js)